[MM-836]: moved the mockSetup function to a new file
