+++ calendar/engine/calendar_test.go
@@ -8,13 +8,8 @@ import (
 	"github.com/golang/mock/gomock"
 	"github.com/pkg/errors"
 
-	"github.com/mattermost/mattermost-plugin-mscalendar/calendar/config"
-	"github.com/mattermost/mattermost-plugin-mscalendar/calendar/engine/mock_plugin_api"
 	"github.com/mattermost/mattermost-plugin-mscalendar/calendar/remote"
-	"github.com/mattermost/mattermost-plugin-mscalendar/calendar/remote/mock_remote"
 	"github.com/mattermost/mattermost-plugin-mscalendar/calendar/store"
-	"github.com/mattermost/mattermost-plugin-mscalendar/calendar/store/mock_store"
-	"github.com/mattermost/mattermost-plugin-mscalendar/calendar/utils/bot/mock_bot"
 
 	"github.com/mattermost/mattermost/server/public/model"
 	"github.com/stretchr/testify/require"
@@ -241,19 +236,19 @@ func TestCreateEvent(t *testing.T) {
 	mscalendar, mockStore, mockPoster, _, mockPluginAPI, mockClient, mockLogger := MockSetup(t)
 
 	tests := []struct {
-		name              string
-		user              *User
-		event             *remote.Event
-		setupMock         func()
-		assertions        func(t *testing.T, createdEvent *remote.Event, err error)
-		expectedEvent     *remote.Event
+		name          string
+		user          *User
+		event         *remote.Event
+		setupMock     func()
+		assertions    func(t *testing.T, createdEvent *remote.Event, err error)
+		expectedEvent *remote.Event
 	}{
 		{
 			name: "error expanding user",
 			user: &User{
 				MattermostUserID: "testMMUserID",
 			},
-			event:             &remote.Event{Subject: "Test Event"},
+			event: &remote.Event{Subject: "Test Event"},
 			setupMock: func() {
 				mockStore.EXPECT().LoadUser("testMMUserID").Return(nil, errors.New("error loading the user")).Times(1)
 			},
@@ -268,7 +263,7 @@ func TestCreateEvent(t *testing.T) {
 				User:             &store.User{Remote: &remote.User{ID: "testRemoteUserID"}},
 				MattermostUserID: "testMMUserID",
 			},
-			event:             &remote.Event{Subject: "Test Event"},
+			event: &remote.Event{Subject: "Test Event"},
 			setupMock: func() {
 				mockStore.EXPECT().LoadUser("testMMUserID").Return(nil, errors.New("not found")).Times(1)
 				mockPluginAPI.EXPECT().GetMattermostUser("testMMUserID")
@@ -288,7 +283,7 @@ func TestCreateEvent(t *testing.T) {
 				User:             &store.User{Remote: &remote.User{ID: "testRemoteUserID"}},
 				MattermostUserID: "testMMUserID",
 			},
-			event:             &remote.Event{Subject: "Test Event"},
+			event: &remote.Event{Subject: "Test Event"},
 			setupMock: func() {
 				mockStore.EXPECT().LoadUser("testMMUserID").Return(nil, errors.New("not found")).Times(1)
 				mockPluginAPI.EXPECT().GetMattermostUser("testMMUserID")
@@ -523,42 +518,4 @@ func TestGetCalendars(t *testing.T) {
 			tt.assertions(t, err, calendars)
 		})
 	}
-}
-
-// revive:disable:unexported-return
-func MockSetup(t *testing.T) (*mscalendar, *mock_store.MockStore, *mock_bot.MockPoster, *mock_remote.MockRemote, *mock_plugin_api.MockPluginAPI, *mock_remote.MockClient, *mock_bot.MockLogger) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockStore := mock_store.NewMockStore(ctrl)
-	mockPoster := mock_bot.NewMockPoster(ctrl)
-	mockRemote := mock_remote.NewMockRemote(ctrl)
-	mockPluginAPI := mock_plugin_api.NewMockPluginAPI(ctrl)
-	mockClient := mock_remote.NewMockClient(ctrl)
-	mockLogger := mock_bot.NewMockLogger(ctrl)
-
-	env := Env{
-		Dependencies: &Dependencies{
-			Store:     mockStore,
-			Poster:    mockPoster,
-			Remote:    mockRemote,
-			PluginAPI: mockPluginAPI,
-			Logger:    mockLogger,
-		},
-	}
-
-	mscalendar := &mscalendar{
-		Env:    env,
-		client: mockClient,
-	}
-
-	mscalendar.Config = &config.Config{
-		Provider: config.ProviderConfig{
-			DisplayName:    "testDisplayName",
-			CommandTrigger: "testCommandTrigger",
-		},
-		PluginVersion: "1.0.0",
-	}
-
-	return mscalendar, mockStore, mockPoster, mockRemote, mockPluginAPI, mockClient, mockLogger
 }
